requirements.txt
pillow>=9.5.0
pycryptodomex>=3.18.0
